% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.tsreg.R
\name{plot.tsreg}
\alias{plot.tsreg}
\title{Plot tsreg object}
\usage{
plot.tsreg(x, ylabs = c("data", "|MOSUM|"), legend = TRUE)
}
\arguments{
\item{x}{Object of class ``tsreg"}

\item{ylabs}{Character. Vector of length 2 with the y-labels on the respective data and MOSUM plots}

\item{legend}{Logical. Include a legend in the plot?}
}
\description{
Plotting function for objects returned by \code{\link{tsreg}}
}
\author{
Ben DeVries
}

