% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/regSpatial.R
\name{regSpatial}
\alias{regSpatial}
\title{Monitor post-disturbance regrowth}
\usage{
regSpatial(x, change, h = 0.5, ..., filename = NULL, mc.cores = 1)
}
\arguments{
\item{x}{RasterBrick. Input time series brick.}

\item{change}{RasterLayer. Raster layer with inistial disturbance dates (e.g. breakpoints from \code{\link{bfastmonitor}}).}

\item{h}{MOSUM bandwidth (see \code{\link{tsreg}})}

\item{...}{Additional arguments to be passed to \code{\link{tsreg}}}

\item{filename}{Optional}

\item{mc.cores}{Numeric. Number of cores for parallel processing (see \code{\link{mc.calc}})}
}
\description{
Monitor post-disturbance regrowth using a LTS rasterBrick
}
\examples{
## ...coming soon...
}
\author{
Ben DeVries
}
\seealso{
tsreg
}

