% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stableHistory.R
\name{stableHistory}
\alias{stableHistory}
\title{Stable history period selection}
\usage{
stableHistory(data_tspp, start, formula, history, level, hpc)
}
\arguments{
\item{data_tspp}{data.frame}

\item{start}{Numeric}

\item{formula}{}

\item{history}{}

\item{level}{}

\item{hpc}{}
}
\description{
Determine stable history from a \code{bfastpp} data.frame. See \code{\link{bfastmonitor}} for more information.
}
\details{
This function is used to determine the stable history in \code{\link{tsreg}}
}
\author{
Jan Verbesselt, Achim Zeileis
}
\seealso{
bfastmonitor
}

