% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tsreg.R
\name{tsreg}
\alias{tsreg}
\title{Monitor post-disturbance regrowth}
\usage{
tsreg(x, change, h, dates = NULL, type = c("irregular", "16-day"),
  startOffset = 0, formula = response ~ harmon, order = 1,
  history = "BP", level = 0.05, w = 3, s = 1, plot = FALSE,
  ylabs = c("data", "|MOSUM|"))
}
\arguments{
\item{x}{Time series. Can be an object of type \code{numeric} (vector), \code{zoo} or \code{ts}}

\item{change}{Numeric. Date of original disturbance. See \code{\link{bfastmonitor}} for a potential method to determine this on a time series.}

\item{dates}{Date. Vector with length \code{length(x)} with observation dates. If omitted, \code{x} will simply be an ordered series with arbitrary (regularly spaced) dates}

\item{startOffset}{Numeric. Number of (decimal) years by which to offset the start of the regrowth monitoring process. Can also be set to "floor", in which case the \code{floor} of the change date will be taken as the start of the monitoring period.}

\item{formula}{Formula to be fit to stable history period (see \code{\link{bfastmonitor}})}

\item{order}{Numeric. Order of the harmonic term of \code{formula} (see \code{\link{bfastmonitor}})}

\item{history}{Method or pre-determined period to use for history model (see \code{\link{bfastmonitor}})}

\item{level}{see \code{\link{bfastmonitor}}}

\item{w}{Numeric. Number of years before a regrowth flag is allowed to be assigned.}

\item{plot}{Logical. Plot the results?}

\item{ylabs}{Character. Vector of length 2 with y labels of the data time series and MOSUM plots, respectively}
}
\value{
object of class \code{tsreg}
}
\description{
Structural monitoring method for post-disturbance forest regrowth in a time series
}
\examples{
data(ndmi)
plot(ndmi, type='b', cex=0.5)

# breakpoint detection with bfastmonitor
x <- bfastts(ndmi, time(ndmi), type="irregular")
bfm <- bfastmonitor(x, start=c(2005, 1), formula=response~harmon, order=1, plot=TRUE)

reg <- tsreg(ndmi, change = bfm$breakpoint, h = 0.5, plot = TRUE)
print(reg)

reg2 <- tsreg(ndmi, change = bfm$breakpoint, startOffset = "floor", h = 0.5, plot = TRUE)
print(reg2)

reg3 <- tsreg(ndmi, change = bfm$breakpoint, startOffset = "floor", h = 0.5, history='all', plot=TRUE)
print(reg3)

reg4 <- tsreg(ndmi, change = bfm$breakpoint, startOffset = "floor", h = 0.5, history='all', s=0, plot=TRUE)
print(reg4)
}
\author{
Ben DeVries
}
\seealso{
regSpatial
}

